requirements.txt
python-dotenv
python-telegram-bot
openai
pandas
SQLAlchemy
openpyxl
